+++ crates/fluvio-cli/src/error.rs
@@ -1,6 +1,7 @@
 use std::io::{Error as IoError, ErrorKind};
 
 use semver::Version;
+use handlebars::TemplateError;
 use fluvio::FluvioError;
 
 #[cfg(feature = "k8s")]
@@ -20,6 +21,8 @@ pub enum CliError {
     IoError(#[from] IoError),
     #[error(transparent)]
     OutputError(#[from] OutputError),
+    #[error("Failed to parse format string")]
+    TemplateError(#[from] TemplateError),
 
     #[cfg(feature = "k8s")]
     #[error("Fluvio cluster error")]
@@ -101,6 +104,10 @@ impl CliError {
                 }
                 ApiError::Code(ErrorCode::TopicNotFound, _) => {
                     println!("Topic not found");
+                    Ok(())
+                }
+                ApiError::Code(ErrorCode::TopicInvalidName, _) => {
+                    println!("Invalid topic name: topic name may only include lowercase letters (a-z), numbers (0-9), and hyphens (-).");
                     Ok(())
                 }
                 _ => Err(self),