+++ crates/fluvio-sc/src/core/mod.rs
@@ -1,3 +1,2 @@
 mod context;
-pub mod common;
 pub use self::context::*;