+++ crates/fluvio/src/admin.rs
@@ -108,20 +108,27 @@ impl FluvioAdmin {
     /// ```
     #[instrument(skip(config))]
     pub async fn connect_with_config(config: &FluvioConfig) -> Result<Self, FluvioError> {
+        use fluvio_sc_schema::objects::WatchRequest;
+        use fluvio_protocol::api::Request;
+
         let connector = DomainConnector::try_from(config.tls.clone())?;
-        let config = ClientConfig::new(&config.endpoint, connector);
+        let config = ClientConfig::new(&config.endpoint, connector, config.use_spu_local_address);
         let inner_client = config.connect().await?;
-        debug!("connected to cluster at: {}", inner_client.config().addr());
+        debug!(addr = %inner_client.config().addr(), "connected to cluster");
 
         let (socket, config, versions) = inner_client.split();
-        let socket = MultiplexerSocket::shared(socket);
-        let metadata = MetadataStores::start(socket.clone()).await?;
-        let versioned_socket = VersionedSerialSocket::new(socket, config, versions);
+        if let Some(watch_version) = versions.lookup_version(WatchRequest::API_KEY) {
+            let socket = MultiplexerSocket::shared(socket);
+            let metadata = MetadataStores::start(socket.clone(), watch_version).await?;
+            let versioned_socket = VersionedSerialSocket::new(socket, config, versions);
 
-        Ok(Self {
-            socket: versioned_socket,
-            metadata,
-        })
+            Ok(Self {
+                socket: versioned_socket,
+                metadata,
+            })
+        } else {
+            Err(FluvioError::Other("WatchApi versio not found".to_string()))
+        }
     }
 
     #[instrument(skip(self, request))]